refactor: extract map logic into reusable mixin
